third_party/blink/web_tests/platform/generic/fast/writing-mode/relative-positioning-percentages-expected.txt

This is a test that relative position percentages are computed with respect to physical dimensions.
PASS: The block is in the correct position.
